GBCore

Documentation for GBCore.

GBCore.AbstractGB

GBCore.CV

GBCore.Fit

GBCore.Genomes

GBCore.Phenomes

GBCore.SimulatedEffects

GBCore.TEBV

GBCore.Trials

Base.:==

Base.:==

Base.:==

Base.:==

Base.:==

Base.:==

Base.:==

Base.filter

Base.filter

Base.filter

Base.hash

Base.hash

Base.hash

Base.hash

Base.hash

Base.hash

Base.hash

Base.merge

Base.merge

Base.merge

Base.sum

GBCore.addcompositetrait

GBCore.addcompositetrait

GBCore.analyse

GBCore.analyse

GBCore.checkdims

GBCore.checkdims

GBCore.checkdims

GBCore.checkdims

GBCore.checkdims

GBCore.checkdims

GBCore.checkdims

GBCore.clone

GBCore.clone

GBCore.clone

GBCore.clone

GBCore.clone

GBCore.clone

GBCore.countlevels

GBCore.dimensions

GBCore.dimensions

GBCore.dimensions

GBCore.dimensions

GBCore.distances

GBCore.distances

GBCore.extractphenomes

GBCore.extractphenomes

GBCore.loci

GBCore.loci_alleles

GBCore.plot

GBCore.plot

GBCore.plot

GBCore.plot

GBCore.plot

GBCore.simulateeffects

GBCore.simulategenomes

GBCore.simulategenomiceffects

GBCore.simulatetrials

GBCore.slice

GBCore.slice

GBCore.slice

GBCore.summarise

GBCore.tabularise

GBCore.tabularise

GBCore.tabularise

GBCore.tabularise

GBCore.trialsmodelsfomulae!

GBCore.@string2formula

GBCore.@stringevaluation

GBCore.AbstractGB — Type

Abstract super type for all GenomicBreeding.jl custom types

GBCore.CV — Type

Cross-validation struct

Contains genomic prediction cross-validation details.

Fields

replication: replication name

fold: fold name

fit: genomic prediction model fit on the training set

validation_populations: vector of validation populations corresponding to each validation entry

validation_entries: corresponding vector of entries in the validation population/s

validation_y_true: corresponding vector of observed phenotypes in the validation population/s

validation_y_pred: corresponding vector of predicted phenotypes in the validation population/s

metrics: dictionary of genomic prediction accuracy metrics on the validation population/s

Constructor

Uses the default contructor.

GBCore.Fit — Type

Genomic prediction model fit

Contains genomic prediction model fit details.

Fields

model: name of the genomic prediction model used

b_hat_labels: names of the loci-alleles used

b_hat: effects of the loci-alleles

trait: name of the trait

entries: names of the entries used in the current cross-validation replication and fold

populations: names of the populations used in the current cross-validation replication and fold

y_true: corresponding observed phenotype values

y_pred: corresponding predicted phenotype values

metrics: dictionary of genomic prediction accuracy metrics, inluding Pearson's correlation, mean absolute error and root mean-squared error

Constructor

Fit(; n=1, l=10)

GBCore.Genomes — Type

Genomes struct

Containes unique entries and loci_alleles where allele frequencies can have missing values

Constructor

Genomes(; n::Int64 = 1, p::Int64 = 2)

Fields

entries: names of the n entries or samples

populations: name/s of the population/s each entries or samples belong to

loci_alleles: names of the p loci-alleles combinations (p = l loci x a-1 alleles) including the

chromsome or scaffold name, position, all alleles, and current allele separated by tabs ("\t")

allele_frequencies: n x p matrix of allele frequencies between 0 and 1 which can have missing values

mask: n x p matrix of boolean mask for selective analyses and slicing

Examples

julia> genomes = Genomes(n=2, p=2)
Genomes(["", ""], ["", ""], ["", ""], Union{Missing, Float64}[missing missing; missing missing], Bool[1 1; 1 1])

julia> fieldnames(Genomes)
(:entries, :populations, :loci_alleles, :allele_frequencies, :mask)

julia> genomes.entries = ["entry_1", "entry_2"];

julia> genomes.populations = ["pop_1", "pop_1"];

julia> genomes.loci_alleles = ["chr1\t12345\tA|T\tA", "chr2\t678910\tC|D\tD"];

julia> genomes.allele_frequencies = [0.5 0.25; 0.9 missing];

julia> genomes.mask = [true true; true false];

julia> genomes
Genomes(["entry_1", "entry_2"], ["pop_1", "pop_1"], ["chr1\t12345\tA|T\tA", "chr2\t678910\tC|D\tD"], Union{Missing, Float64}[0.5 0.25; 0.9 missing], Bool[1 1; 1 0])

GBCore.Phenomes — Type

Phenomes struct

Constains unique entries and traits where phenotype data can have missing values

Constructor

Phenomes(; n::Int64 = 1, t::Int64 = 2)

Fields

entries: names of the n entries or samples

populations: name/s of the population/s each entries or samples belong to

traits: names of the t traits

phenotypes: n x t matrix of numeric (R) phenotype data which can have missing values

mask: n x t matrix of boolean mask for selective analyses and slicing

Examples

julia> phenomes = Phenomes(n=2, t=2)
Phenomes(["", ""], ["", ""], ["", ""], Union{Missing, Float64}[missing missing; missing missing], Bool[1 1; 1 1])

julia> fieldnames(Phenomes)
(:entries, :populations, :traits, :phenotypes, :mask)

julia> phenomes.entries = ["entry_1", "entry_2"];

julia> phenomes.populations = ["pop_A", "pop_B"];

julia> phenomes.traits = ["height", "yield"];

julia> phenomes.phenotypes = [200.0 2.5; 150.0 missing];

julia> phenomes.mask = [true true; true false];

julia> phenomes
Phenomes(["entry_1", "entry_2"], ["pop_A", "pop_B"], ["height", "yield"], Union{Missing, Float64}[200.0 2.5; 150.0 missing], Bool[1 1; 1 0])

GBCore.SimulatedEffects — Type

SimulatedEffects struct

Contains:

Identification, i.e. trait, year, season, harvest, site, and replication

Additive environmental effects:

year

season

site

Environmental interaction effects:

seasonsxyear

harvestsxseasonxyear

sitesxharvestxseasonxyear

Spatial effects including the field layout per year-season-harvest-site combination

field_layout

replicationsxsitexharvestxseasonxyear

blocksxsitexharvestxseasonxyear

rowsxsitexharvestxseasonxyear

colsxsitexharvestxseasonxyear

Genetic effects

additive_genetic

dominance_genetic

epistasis_genetic

GxE effects

additiveallelexsitexharvestxseasonx_year

dominanceallelexsitexharvestxseasonx_year

epistasisallelexsitexharvestxseasonx_year

Constructor

SimulatedEffects()

GBCore.TEBV — Type

Trial-estimate breeding values (TEBV) struct

Contains trial-estimated breeding values as generated by analyse(trials::Trials, ...).

Fields

traits: names of the traits t traits

formulae: best-fitting formula for each trait

models: best-fitting linear mixed model for each trait

df_BLUEs: vector of data frames of best linear unbiased estimators or fixed effects table of each best fitting model

df_BLUPs: vector of data frames of best linear unbiased predictors or random effects table of each best fitting model

phenomes: vector of Phenomes structs each containing the breeding values

Examples

julia> tebv = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], phenomes=[Phenomes(n=1, t=1)]);

julia> tebv.traits
1-element Vector{String}:
 ""

GBCore.Trials — Type

Trials struct

Contains phenotype data across years, seasons, harvest, sites, populations, replications, blocks, rows, and columns

Constructor

Trials(; n::Int64 = 2, p::Int64 = 2)

Fields

phenotypes: n x t matrix of numeric phenotype data which can have missing values

traits: names of the traits t traits

years: names of the years corresponding to each row in the phenotype matrix

seasons: names of the seasons corresponding to each row in the phenotype matrix

harvests: names of the harvests corresponding to each row in the phenotype matrix

sites: names of the sites corresponding to each row in the phenotype matrix

replications: names of the replications corresponding to each row in the phenotype matrix

blocks: names of the blocks corresponding to each row in the phenotype matrix

rows: names of the rows corresponding to each row in the phenotype matrix

cols: names of the cols corresponding to each row in the phenotype matrix

entries: names of the entries corresponding to each row in the phenotype matrix

populations: names of the populations corresponding to each row in the phenotype matrix

Examples

julia> trials = Trials(n=1, t=2)
Trials(Union{Missing, Float64}[missing missing], ["", ""], [""], [""], [""], [""], [""], [""], [""], [""], [""], [""])

julia> fieldnames(Trials)
(:phenotypes, :traits, :years, :seasons, :harvests, :sites, :replications, :blocks, :rows, :cols, :entries, :populations)

Base.:== — Method

Base.:(==)(x::CV, y::CV)::Bool

Equality of CV structs using the hash function defined for CV structs.

Examples

julia> fit = Fit(n=1, l=2);

julia> cv_1 = CV("replication_1", "fold_1", fit, ["population_1"], ["entry_1"], [0.0], [0.0], fit.metrics);

julia> cv_2 = clone(cv_1);

julia> cv_3 = clone(cv_1); cv_3.replication = "other_replication";

julia> cv_1 == cv_2
true

julia> cv_1 == cv_3
false

Base.:== — Method

Base.:(==)(x::Fit, y::Fit)::Bool

Equality of Fit structs using the hash function defined for Fit structs.

Examples

julia> fit_1 = Fit(n=1, l=4);

julia> fit_2 = Fit(n=1, l=4);

julia> fit_3 = Fit(n=1, l=2);

julia> fit_1 == fit_2
true

julia> fit_1 == fit_3
false

Base.:== — Method

Base.:(==)(x::Genomes, y::Genomes)::Bool

Equality of Genomes structs using the hash function defined for Genomes structs.

Examples

julia> genomes_1 = genomes = Genomes(n=2,p=4);

julia> genomes_2 = genomes = Genomes(n=2,p=4);

julia> genomes_3 = genomes = Genomes(n=1,p=2);

julia> genomes_1 == genomes_2
true

julia> genomes_1 == genomes_3
false

Base.:== — Method

Base.:(==)(x::Phenomes, y::Phenomes)::Bool

Equality of Phenomes structs using the hash function defined for Phenomes structs.

Examples

julia> phenomes_1 = phenomes = Phenomes(n=2, t=4);

julia> phenomes_2 = phenomes = Phenomes(n=2, t=4);

julia> phenomes_3 = phenomes = Phenomes(n=1, t=2);

julia> phenomes_1 == phenomes_2
true

julia> phenomes_1 == phenomes_3
false

Base.:== — Method

Base.:(==)(x::SimulatedEffects, y::SimulatedEffects)::Bool

Equality of SimulatedEffects structs using the hash function defined for SimulatedEffects structs.

Examples

julia> effects_1 = SimulatedEffects();

julia> effects_2 = SimulatedEffects();

julia> effects_3 = SimulatedEffects(); effects_3.id[1] = "SOMETHING_ELSE";

julia> effects_1 == effects_2
true

julia> effects_1 == effects_3
false

Base.:== — Method

Base.:(==)(x::TEBV, y::TEBV)::Bool

Equality of TEBV structs using the hash function defined for TEBV structs.

Examples

julia> tebv_1 = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], phenomes=[Phenomes(n=1,t=1)]);

julia> tebv_2 = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], phenomes=[Phenomes(n=1,t=1)]);

julia> tebv_3 = TEBV(traits=["SOMETHING_ELSE"], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], phenomes=[Phenomes(n=1,t=1)]);

julia> tebv_1 == tebv_2
true

julia> tebv_1 == tebv_3
false

Base.:== — Method

Base.:(==)(x::Trials, y::Trials)::Bool

Equality of Trials structs using the hash function defined for Trials structs.

Examples

julia> trials_1 = trials = Trials(n=2, t=4);

julia> trials_2 = trials = Trials(n=2, t=4);

julia> trials_3 = trials = Trials(n=1, t=2);

julia> trials_1 == trials_2
true

julia> trials_1 == trials_3
false

Base.filter — Method

filter(
    genomes::Genomes;
    maf::Float64,
    max_entry_sparsity::Float64 = 0.0,
    max_locus_sparsity::Float64 = 0.0,
    chr_pos_allele_ids::Union{Missing,Vector{String}} = missing,
)::Genomes

Filter a Genomes struct by minimum allele frequency, entries sparsity, loci sparsity, and/or vector of locus-allele names

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> filtered_genomes_1 = filter(genomes, 0.1);

julia> filtered_genomes_2 = filter(genomes, 0.1, chr_pos_allele_ids=genomes.loci_alleles[1:1000]);

julia> size(genomes.allele_frequencies)
(100, 3000)

julia> size(filtered_genomes_1.allele_frequencies)
(100, 1236)

julia> size(filtered_genomes_2.allele_frequencies)
(100, 388)

Base.filter — Method

filter(genomes::Genomes)::Genomes

Filter a Genomes struct using its mask matrix where all rows and columns with at least one false value are excluded

Examples

julia> genomes = simulategenomes(verbose=false); genomes.mask[1:10, 42:100] .= false;
    
julia> filtered_genomes = filter(genomes);

julia> size(filtered_genomes.allele_frequencies)
(90, 9941)

Base.filter — Method

filter(phenomes::Phenomes)::Phenomes

Filter a Phenomes struct using its mask matrix where all rows and columns with at least one false value are excluded

Examples

julia> phenomes = Phenomes(n=10, t=3); phenomes.entries = string.("entry_", 1:10); phenomes.populations .= "pop_1"; phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = fill(0.0, 10,3);

julia> phenomes.mask .= true; phenomes.mask[6:10, 1] .= false;
    
julia> filtered_phenomes = filter(phenomes);

julia> size(filtered_phenomes.phenotypes)
(5, 2)

Base.hash — Method

Base.hash(x::CV, h::UInt)::UInt

Hash a CV struct using the entries, populations and locialleles. We deliberately excluded the allelefrequencies, and mask for efficiency.

Examples

julia> fit = Fit(n=1, l=2);

julia> cv = CV("replication_1", "fold_1", fit, ["population_1"], ["entry_1"], [0.0], [0.0], fit.metrics);

julia> typeof(hash(cv))
UInt64

Base.hash — Method

Base.hash(x::Fit, h::UInt)::UInt

Hash a Fit struct using the entries, populations and locialleles. We deliberately excluded the allelefrequencies, and mask for efficiency.

Examples

julia> fit = Fit(n=1, l=2);

julia> typeof(hash(fit))
UInt64

Base.hash — Method

Base.hash(x::Genomes, h::UInt)::UInt

Hash a Genomes struct using the entries, populations and locialleles. We deliberately excluded the allelefrequencies, and mask for efficiency.

Examples

julia> genomes = Genomes(n=2, p=2);

julia> typeof(hash(genomes))
UInt64

Base.hash — Method

Base.hash(x::Phenomes, h::UInt)::UInt

Hash a Phenomes struct.

Examples

julia> phenomes = Phenomes(n=2, t=2);

julia> typeof(hash(phenomes))
UInt64

Base.hash — Method

Base.hash(x::SimulatedEffects, h::UInt)::UInt

Hash a SimulatedEffects struct.

Examples

julia> effects = SimulatedEffects();

julia> typeof(hash(effects))
UInt64

Base.hash — Method

Base.hash(x::TEBV, h::UInt)::UInt

Hash a TEBV struct using the traits, formualae and phenomes. We deliberately excluded the models, dfBLUEs, and dfBLUPs for efficiency.

Examples

julia> tebv = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], phenomes=[Phenomes(n=1,t=1)]);

julia> typeof(hash(tebv))
UInt64

Base.hash — Method

Base.hash(x::Trials, h::UInt)::UInt

Hash a Trials struct.

Examples

julia> trials = Trials(n=2, t=2);

julia> typeof(hash(trials))
UInt64

Base.merge — Method

merge(
    genomes::Genomes,
    other::Genomes;
    conflict_resolution::Tuple{Float64,Float64} = (0.5, 0.5),
    verbose::Bool = true,
)::Genomes

Merge two Genomes structs using a tuple of conflict resolution weights

Examples

julia> n = 100; l = 5_000; n_alleles = 2;

julia> all = simulategenomes(n=n, l=l, n_alleles=n_alleles, verbose=false);

julia> genomes = slice(all, idx_entries=collect(1:Int(floor(n*0.75))), idx_loci_alleles=collect(1:Int(floor(l*(n_alleles-1)*0.75))));

julia> other = slice(all, idx_entries=collect(Int(floor(n*0.50)):n), idx_loci_alleles=collect(Int(floor(l*(n_alleles-1)*0.50)):l*(n_alleles-1)));

julia> merged_genomes = merge(genomes, other, conflict_resolution=(0.75, 0.25), verbose=false);

julia> size(merged_genomes.allele_frequencies)
(100, 5000)

julia> sum(ismissing.(merged_genomes.allele_frequencies))
123725

Base.merge — Method

merge(genomes::Genomes, phenomes::Phenomes; keep_all::Bool=true)::Tuple{Genomes,Phenomes}

Merge a Genomes struct with a Phenomes struct using union or intersection

Examples

julia> genomes = simulategenomes(n=10, verbose=false);

julia> trials, effects = simulatetrials(genomes=slice(genomes, idx_entries=collect(1:5), idx_loci_alleles=collect(1:length(genomes.loci_alleles))), f_add_dom_epi=[0.90 0.05 0.05;], n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=2, verbose=false);

julia> phenomes = Phenomes(n=5, t=1);

julia> phenomes.entries = trials.entries[1:5];

julia> phenomes.populations = trials.populations[1:5];

julia> phenomes.traits = trials.traits;

julia> phenomes.phenotypes = trials.phenotypes[1:5, :];

julia> phenomes.mask .= true;

julia> genomes_merged_1, phenomes_merged_1 = merge(genomes, phenomes, keep_all=true);

julia> size(genomes_merged_1.allele_frequencies), size(phenomes_merged_1.phenotypes)
((10, 10000), (10, 1))

julia> genomes_merged_2, phenomes_merged_2 = merge(genomes, phenomes, keep_all=false);

julia> size(genomes_merged_2.allele_frequencies), size(phenomes_merged_2.phenotypes)
((5, 10000), (5, 1))

Base.merge — Method

merge(
    phenomes::Phenomes,
    other::Phenomes;
    conflict_resolution::Tuple{Float64,Float64} = (0.5, 0.5),
    verbose::Bool = true,
)::Phenomes

Merge two Phenomes structs using a tuple of conflict resolution weights

Examples

julia> all = Phenomes(n=10, t=3); all.entries = string.("entry_", 1:10); all.traits = ["A", "B", "C"]; all.phenotypes = rand(10,3);

julia> phenomes = slice(all, idx_entries=collect(1:7), idx_traits=[1,2]);

julia> other = slice(all, idx_entries=collect(5:10), idx_traits=[2,3]);

julia> merged_phenomes = merge(phenomes, other, conflict_resolution=(0.75, 0.25), verbose=false);

julia> size(merged_phenomes.phenotypes)
(10, 3)

julia> sum(ismissing.(merged_phenomes.phenotypes))
7

Base.sum — Method

sum(effects::SimulatedEffects)::Tuple{Int64, Int64, Int64}

Sum up the simulated effects to generate the simulated phenotype values

Examples

julia> effects = SimulatedEffects();

julia> sum(effects)
1-element Vector{Float64}:
 0.0

julia> effects.additive_genetic[1] = pi;

julia> sum(effects)
1-element Vector{Float64}:
 3.141592653589793

GBCore.addcompositetrait — Method

addcompositetrait(phenomes::Phenomes; composite_trait_name::String, formula_string::String)::Phenomes

Add a composite trait from existing traits in a Phenomes struct

julia> phenomes = Phenomes(n=10, t=3); phenomes.entries = string.("entry_", 1:10); phenomes.populations .= "pop_1"; phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = rand(10,3);

julia> phenomes_new = addcompositetrait(phenomes, composite_trait_name = "some_wild_composite_trait", formula_string = "A");

julia> phenomes_new.phenotypes[:, end] == phenomes.phenotypes[:, 1]
true

julia> phenomes_new = addcompositetrait(phenomes, composite_trait_name = "some_wild_composite_trait", formula_string = "(A^B) + (C/A) - sqrt(abs(B-A)) + log(1.00 + C)");

julia> phenomes_new.phenotypes[:, end] == (phenomes.phenotypes[:,1].^phenomes.phenotypes[:,2]) .+ (phenomes.phenotypes[:,3]./phenomes.phenotypes[:,1]) .- sqrt.(abs.(phenomes.phenotypes[:,2].-phenomes.phenotypes[:,1])) .+ log.(1.00 .+ phenomes.phenotypes[:,3])
true

GBCore.addcompositetrait — Method

addcompositetrait(trials::Trials; composite_trait_name::String, formula_string::String)::Trials

Add a composite trait from existing traits in a Trials struct

julia> trials, _ = simulatetrials(genomes = simulategenomes(verbose=false), verbose=false);

julia> trials_new = addcompositetrait(trials, composite_trait_name = "some_wild_composite_trait", formula_string = "trait_1");

julia> trials_new.phenotypes[:, end] == trials.phenotypes[:, 1]
true

julia> trials_new = addcompositetrait(trials, composite_trait_name = "some_wild_composite_trait", formula_string = "(trait_1^(trait_2/100)) + (trait_3/trait_1) - sqrt(abs(trait_2-trait_1)) + log(1.00 + trait_3)");

julia> trials_new.phenotypes[:, end] == (trials.phenotypes[:,1].^(trials.phenotypes[:,2]/100)) .+ (trials.phenotypes[:,3]./trials.phenotypes[:,1]) .- sqrt.(abs.(trials.phenotypes[:,2].-trials.phenotypes[:,1])) .+ log.(1.00 .+ trials.phenotypes[:,3])
true

GBCore.analyse — Function

analyse(
    trials::Trials,
    formula_string::String = "";
    traits::Union{Nothing,Vector{String}} = nothing,
    max_levels::Int64 = 100,
    max_time_per_model::Int64 = 60,
    covariates_continuous::Union{Nothing,Vector{String}} = nothing,
    verbose::Bool = true,
)::TEBV

Analyse trials

Arguments

trials: Trials struct

max_levels: maximum number of non-entry factor levels to include in the linear mixed models (default = 100)

max_time_per_model: maximum time in seconds for fitting each linear mixed model (default = 60)

verbose: Show trials analysis progress bar? (default = true)

Outputs

TEBV struct containing the trait names, the best fitting formulae, models, BLUEs, and BLUPs for each trait

Examples

julia> trials, _simulated_effects = simulatetrials(genomes = simulategenomes(n=10, verbose=false), n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=10, verbose=false);

julia> tebv_1 = analyse(trials, "trait_1 ~ 1 + (1|entries)", max_levels=50, verbose=false);

julia> tebv_1.traits
3-element Vector{String}:
 "trait_1"
 "trait_2"
 "trait_3"

julia> tebv_2 = analyse(trials, max_levels=50, verbose=false);

julia> mean(tebv_2.phenomes[1].phenotypes) < mean(tebv_2.phenomes[2].phenotypes)
true

julia> trials = addcompositetrait(trials, composite_trait_name = "covariate", formula_string = "(trait_1 + trait_2) / (trait_3 + 0.0001)");

julia> tebv_3 = Suppressor.@suppress analyse(trials, "y ~ 1 + covariate + entries + (1|blocks)", max_levels=50, verbose=false);

julia> mean(tebv_3.phenomes[1].phenotypes) < mean(tebv_3.phenomes[2].phenotypes)
true

julia> tebv_4 = Suppressor.@suppress analyse(trials, max_levels=50, covariates_continuous=["covariate"], verbose=false);

julia> mean(tebv_4.phenomes[1].phenotypes) < mean(tebv_4.phenomes[2].phenotypes)
true

GBCore.analyse — Method

analyse(
    df::DataFrame; 
    formulae::Vector{String},
    idx_parallel_models::Vector{Int64},
    idx_iterative_models::Vector{Int64},
    max_time_per_model::Int64 = 60,
    verbose::Bool=false)::Tuple{String, Any, DataFrame, DataFrame, Phenomes}

Fit univarite (one trait) linear mixed models to extract the effects of the entries the best fitting model.

We have the following guiding principles:

Avoid over-parameterisation we'll have enough of that with the genomic prediction models

We will fit mixed models with unstructure variance-covariance matrix of the random effects

We prefer REML over ML

We compare BLUEs vs BLUPs of entries

Examples

julia> trials, _ = simulatetrials(genomes = simulategenomes(verbose=false), verbose=false);

julia> df = tabularise(trials);

julia> formulae, n_levels = trialsmodelsfomulae!(df; trait = "trait_1", max_levels = 10);

julia> idx_parallel_models::Vector{Int64} = findall(n_levels .<= (15));

julia> idx_iterative_models::Vector{Int64} = findall((n_levels .<= (15)) .!= true);

julia> formula_string, model, df_BLUEs, df_BLUPs, phenomes = analyse(df, formulae=formulae, idx_parallel_models=idx_parallel_models, idx_iterative_models=idx_iterative_models);

julia> length(phenomes.entries) == length(unique(df.entries))
true

julia> df_2 = df[(df.years .== df.years[1]) .&& (df.harvests .== df.harvests[1]) .&& (df.seasons .== df.seasons[1]) .&& (df.sites .== df.sites[1]) .&& (df.replications .== df.replications[1]), :];

julia> formula_string_2, model_2, df_BLUEs_2, df_BLUPs_2, phenomes_2 = analyse(df_2, formulae=["trait_1 ~ 1 + 1|entries"]);

julia> cor(phenomes_2.phenotypes[sortperm(phenomes_2.entries),1], df_2.trait_1[sortperm(df_2.entries)]) == 1.00
true

GBCore.checkdims — Method

checkdims(cv::CV)::Bool

Check dimension compatibility of the fields of the CV struct

Examples

julia> fit = Fit(n=1, l=2);

julia> cv = CV("replication_1", "fold_1", fit, ["population_1"], ["entry_1"], [0.0], [0.0], fit.metrics);

julia> checkdims(cv)
true

julia> cv.validation_y_true = [0.0, 0.0];

julia> checkdims(cv)
false

GBCore.checkdims — Method

checkdims(fit::Fit)::Bool

Check dimension compatibility of the fields of the Fit struct

Examples

julia> fit = Fit(n=1, l=4);

julia> checkdims(fit)
true

julia> fit.b_hat_labels = ["chr1\t1\tA|T\tA"];

julia> checkdims(fit)
false

GBCore.checkdims — Method

checkdims(genomes::Genomes)::Bool

Check dimension compatibility of the fields of the Genomes struct

Examples

julia> genomes = Genomes(n=2,p=4);

julia> checkdims(genomes)
false

julia> genomes.entries = ["entry_1", "entry_2"];

julia> genomes.loci_alleles = ["chr1\t1\tA|T\tA", "chr1\t2\tC|G\tG", "chr2\t3\tA|T\tA", "chr2\t4\tG|T\tG"];

julia> checkdims(genomes)
true

GBCore.checkdims — Method

checkdims(y::Phenomes)::Bool

Check dimension compatibility of the fields of the Phenomes struct

Examples

julia> y = Phenomes(n=2, t=2);

julia> checkdims(y)
false

julia> y.entries = ["entry_1", "entry_2"];

julia> y.traits = ["trait_1", "trait_2"];

julia> checkdims(y)
true

GBCore.checkdims — Method

checkdims(effects::SimulatedEffects)::Bool

Check dimension compatibility of the fields of the SimulatedEffects struct

Examples

julia> effects = SimulatedEffects();

julia> checkdims(effects)
true

julia> effects.id = ["beaking_change"];

julia> checkdims(effects)
false

GBCore.checkdims — Method

checkdims(y::TEBV)::Bool

Check dimension compatibility of the fields of the TEBV struct

Examples

julia> tebv = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], phenomes=[Phenomes(n=1,t=1)]);

julia> checkdims(tebv)
true

GBCore.checkdims — Method

checkdims(trials::Trials)::Bool

Check dimension compatibility of the fields of the Trials struct

Examples

julia> trials = Trials(n=1, t=2);

julia> trials.entries = ["entry_1"]; trials.traits = ["trait_1", "trait_2"];

julia> checkdims(trials)
true

julia> trials.entries = ["entering_2_entries", "instead_of_just_1"];

julia> checkdims(trials)
false

GBCore.clone — Method

clone(x::CV)::CV

Clone a CV object

Example

julia> fit = Fit(n=1, l=2);

julia> cv = CV("replication_1", "fold_1", fit, ["population_1"], ["entry_1"], [0.0], [0.0], fit.metrics);

julia> copy_cv = clone(cv)
CV("replication_1", "fold_1", Fit("", ["", ""], [0.0, 0.0], "", [""], [""], [0.0], [0.0], Dict("" => 0.0)), ["population_1"], ["entry_1"], [0.0], [0.0], Dict("" => 0.0))

GBCore.clone — Method

clone(x::Fit)::Fit

Clone a Fit object

Example

julia> fit = Fit(n=1, l=2);

julia> copy_fit = clone(fit)
Fit("", ["", ""], [0.0, 0.0], "", [""], [""], [0.0], [0.0], Dict("" => 0.0))

GBCore.clone — Method

clone(x::Genomes)::Genomes

Clone a Genomes object

Example

julia> genomes = Genomes(n=2, p=2);

julia> copy_genomes = clone(genomes)
Genomes(["", ""], ["", ""], ["", ""], Union{Missing, Float64}[missing missing; missing missing], Bool[1 1; 1 1])

GBCore.clone — Method

clone(x::Phenomes)::Phenomes

Clone a Phenomes object

Example

julia> phenomes = Phenomes(n=2, t=2);

julia> copy_phenomes = clone(phenomes)
Phenomes(["", ""], ["", ""], ["", ""], Union{Missing, Float64}[missing missing; missing missing], Bool[1 1; 1 1])

GBCore.clone — Method

clone(x::TEBV)::TEBV

Clone a TEBV object

Example

julia> tebv = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], phenomes=[Phenomes(n=1,t=1)]);

julia> copy_tebv = clone(tebv);

julia> copy_tebv.traits == tebv.traits
true

julia> copy_tebv.phenomes == tebv.phenomes
true

GBCore.clone — Method

clone(x::Trials)::Trials

Clone a Trials object

Example

julia> trials = Trials(n=2, t=2);

julia> copy_trials = clone(trials)
Trials(Union{Missing, Float64}[missing missing; missing missing], ["", ""], ["", ""], ["", ""], ["", ""], ["", ""], ["", ""], ["", ""], ["", ""], ["", ""], ["", ""], ["", ""])

GBCore.countlevels — Method

countlevels(df::DataFrame; column_names::Vector{String})::Int64

Count the total number of factor levels across the column names provided.

GBCore.dimensions — Method

dimensions(genomes::Genomes)::Dict{String, Int64}

Count the number of entries, populations, loci-alleles combination, loci, maximum number of alleles per locus, and number of missing data in the Genomes struct

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> dimensions(genomes)
Dict{String, Int64} with 7 entries:
  "n_entries"      => 100
  "n_chr"          => 7
  "n_loci"         => 1000
  "n_loci_alleles" => 3000
  "n_populations"  => 1
  "n_missing"      => 0
  "max_n_alleles"  => 4

GBCore.dimensions — Method

dimensions(phenomes::Phenomes)::Dict{String, Int64}

Count the number of entries, populations, and traits in the Phenomes struct

Examples

julia> phenomes = Phenomes(n=10, t=3); phenomes.entries = string.("entry_", 1:10); phenomes.populations .= "pop_1"; phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = fill(0.0, 10,3);

julia> dimensions(phenomes)
Dict{String, Int64} with 8 entries:
  "n_total"       => 30
  "n_zeroes"      => 30
  "n_nan"         => 0
  "n_entries"     => 10
  "n_traits"      => 3
  "n_inf"         => 0
  "n_populations" => 1
  "n_missing"     => 0

GBCore.dimensions — Method

dimensions(tebv::TEBV)::Dict{String, Int64}

Count the number of entries, populations, and traits in the TEBV struct

Examples

julia> tebv = TEBV(traits=["trait_1"], formulae=["trait_1 ~ 1 + 1|entries"], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], phenomes=[Phenomes(n=1,t=1)]);

julia> dimensions(tebv)
Dict{String, Int64} with 8 entries:
  "n_total"       => 1
  "n_zeroes"      => 0
  "n_nan"         => 0
  "n_entries"     => 1
  "n_traits"      => 1
  "n_inf"         => 0
  "n_populations" => 1
  "n_missing"     => 1

GBCore.dimensions — Method

dimensions(trials::Trials)::Dict{String, Int64}

Count the number of entries, populations, and traits in the Trials struct

Examples

julia> trials = Trials(n=1, t=2);

julia> trials.entries = ["entry_1"]; trials.traits = ["trait_1", "trait_2"];

julia> dimensions(trials)
Dict{String, Int64} with 16 entries:
  "n_zeroes"       => 0
  "n_harvests"     => 1
  "n_nan"          => 0
  "n_entries"      => 1
  "n_traits"       => 2
  "n_seasons"      => 1
  "n_rows"         => 1
  "n_blocks"       => 1
  "n_missing"      => 2
  "n_inf"          => 0
  "n_total"        => 2
  "n_replications" => 1
  "n_years"        => 1
  "n_sites"        => 1
  "n_cols"         => 1
  "n_populations"  => 1

GBCore.distances — Method

distances(
    genomes::Genomes; 
    distance_metrics::Vector{String}=["euclidean", "correlation", "mad", "rmsd", "χ²"]
    idx_loci_alleles::Union{Nothing, Vector{Int64}} = nothing,
)::Tuple{Vector{String}, Vector{String}, Dict{String, Matrix{Float64}}}

Estimate pairwise distances between loci_alleles and entries. Sparsity leading to less than 2 pairs will yield -Inf values in the resulting matrices. Matrices with how many pairs were used to estimate the distance and correlation matrices are also included as well as the locus-allele and entry names.

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> (loci_alleles, entries, dist) = distances(genomes, distance_metrics=["correlation", "χ²"]);

julia> sort(string.(keys(dist))) == ["entries|correlation", "entries|counts", "entries|χ²", "loci_alleles|correlation", "loci_alleles|counts", "loci_alleles|χ²"]
true

julia> C = dist["entries|correlation"]; C[diagind(C)] == repeat([1], length(genomes.entries))
true

GBCore.distances — Method

distances(
    phenomes::Phenomes; 
    distance_metrics::Vector{String}=["euclidean", "correlation", "mad", "rmsd", "χ²"],
    standardise_traits::Bool = false,
)::Tuple{String, Matrix{Float64}}Tuple{Vector{String}, Vector{String}, Dict{String, Matrix{Float64}}}

Estimate pairwise distances between traits and entries. Sparsity leading to less than 2 pairs will yield -Inf values in the resulting matrices. Matrices with how many pairs were used to estimate the distance and correlation matrices are also included as well as the trait and entry names. Note that prior to computing correlations between entries, the phenotype matrix was standardised, i.e. each trait was transformed to have a mean of 0 and standard deviation of 1.

julia> phenomes = Phenomes(n=10, t=3); phenomes.entries = string.("entry_", 1:10); phenomes.populations .= "pop_1"; phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = rand(10,3); phenomes.phenotypes[2,2] = missing;

julia> (traits, entries, dist) = distances(phenomes, distance_metrics=["correlation", "χ²"]);

julia> sort(string.(keys(dist))) == ["entries|correlation", "entries|counts", "entries|χ²", "traits|correlation", "traits|counts", "traits|χ²"]
true

julia> C = dist["entries|correlation"]; C[diagind(C)] == repeat([1], length(phenomes.entries))
true

julia> dist["traits|counts"][:, 2] == dist["traits|counts"][2, :] == repeat([9], length(phenomes.traits))
true

julia> dist["entries|counts"][:, 2] == dist["entries|counts"][2, :] == repeat([2], length(phenomes.entries))
true

GBCore.extractphenomes — Method

extractphenomes(tebv::TEBV)::Phenomes

Extract Phenomes from TEBV

julia> trials, _simulated_effects = simulatetrials(genomes = simulategenomes(n=10, verbose=false), n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=10, verbose=false);

julia> tebv = analyse(trials, max_levels=50, verbose=false);

julia> phenomes = extractphenomes(tebv);

julia> phenomes.traits == ["trait_1", "trait_2", "trait_3"]
true

GBCore.extractphenomes — Method

extractphenomes(trials::Trials)::Phenomes

Extract Phenomes from Trials struct. Each trait-by-environment variables combination make up the traits in the resulting Phenomes struct. The trait names start with trait name in Trials suffixed by the trait-by-environment variables combinations. If there is a single environment variables combination, then no suffix is added to the trait name.

Examples

julia> trials, _ = simulatetrials(genomes = simulategenomes(verbose=false), verbose=false);

julia> phenomes = extractphenomes(trials);

julia> size(phenomes.phenotypes)
(100, 384)

GBCore.loci — Method

loci(genomes::Genomes)::Tuple{Vector{String},Vector{Int64},Vector{Int64},Vector{Int64}}

Extract chromosome names, positions, start and end indexes of each locus across loci

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> chromosomes, positions, loci_ini_idx, loci_fin_idx = loci(genomes);

julia> length(chromosomes), length(positions), length(loci_ini_idx), length(loci_fin_idx)
(1000, 1000, 1000, 1000)

GBCore.loci_alleles — Method

loci_alleles(genomes::Genomes)::Tuple{Vector{String},Vector{Int64},Vector{String}}

Extract chromosomes, positions, and alleles across loci-allele combinations

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> chromosomes, positions, alleles = loci_alleles(genomes);

julia> length(chromosomes), length(positions), length(alleles)
(3000, 3000, 3000)

GBCore.plot — Function

plot(genomes::Genomes)::Nothing

Plot allele frequencies

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> GBCore.plot(genomes)

GBCore.plot — Function

plot(fit::Fit, distribution::Any=[TDist(1), Normal()][1], α::Float64=0.05)

Manhattan plot

Examples

julia> distribution = [TDist(1), Normal()][2];

julia> fit = Fit(n=100, l=10_000); fit.b_hat = rand(distribution, 10_000);  α = 0.05;

julia> GBCore.plot(fit);

GBCore.plot — Method

plot(phenomes::Phenomes)::Nothing

Plot histogram/s of the trait value/s and a heatmap of trait correlations

Examples

julia> phenomes = Phenomes(n=10, t=3); phenomes.entries = string.("entry_", 1:10); phenomes.populations .= "pop_1"; phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = rand(10,3);

julia> GBCore.plot(phenomes);

GBCore.plot — Method

plot(tebv::TEBV)

Plot TEBV output by plotting the resulting Phenomes struct

GBCore.plot — Method

plot(trials::Trials)::Nothing

Plot histogram/s of the trait value/s and a heatmap of trait correlations across the entire trial. Additionally, plot mean trait values per year, season, harvest, site, replications, row, column, and population for each trait.

Examples

julia> trials, _ = simulatetrials(genomes = simulategenomes(verbose=false), verbose=false);

julia> GBCore.plot(trials);

GBCore.simulateeffects — Method

Simulate effects

Sample p x q effects from a multivariate normal distribution with μ~Exp(λ) and Σ=μμ'

Arguments

p: number of correlated effects to simulate (default = 2)

q: number times to simulate the correlated effects from the same distribution (default = 1)

λ: parameter of the exponential distritbution from which the means will be sampled from (default = 1.00)

seed: randomisation seed (default = 42)

Output

p x q matrix of correlated effects

Examples

julia> θ::Matrix{Float64} = simulateeffects();

julia> sum(abs.(θ - [-0.0886501800782904; -0.596478483888422])) < 0.00001
true

GBCore.simulategenomes — Method

Simulate genomes

Arguments

n: number of entries (default = 100)

l: number of loci (default = 10_000)

n_chroms: number of chromosomes (default = 7)

n_alleles: number of alleles per locus (default = 2)

max_pos: total length of the genome in base-pairs (bp) (default = 135000000)

ld_corr_50perc_kb: distance in bp at which linkage expressed as correlation between a pair of loci is at 50% (default = 1_000)

μ_β_params: the shape parameters of the Beta distribution from which the mean allele frequencies will be sampled (default = (0.5, 0.5); U-shaped distribution; you may use (2.0, 2.0) for a bell-shaped distribution)

sparsity: Proportion of missing data (default = 0.0)

seed: psuedo-random number generator seed for replicability (default = 42)

verbose: Show progress bar and plot the linkage heatmap into an svg file? (default = true)

Output

Genomes

Examples

julia> genomes = simulategenomes(n=100, l=10_000, n_alleles=3, verbose=false);

julia> length(genomes.entries)
100

julia> length(genomes.populations)
100

julia> length(genomes.loci_alleles)
20000

julia> size(genomes.allele_frequencies)
(100, 20000)

julia> mean(ismissing.(genomes.allele_frequencies))
0.0

julia> rng::TaskLocalRNG = Random.seed!(123);

julia> idx = StatsBase.sample(rng, range(1, 20_000, step=2), 250, replace = false, ordered = true);

julia> correlations = StatsBase.cor(genomes.allele_frequencies[:, idx]);

julia> correlations[10,10] == 1.00
true

julia> correlations[10,10] > correlations[10,250]
true

julia> genomes = simulategenomes(n=100, l=10_000, n_alleles=3, sparsity=0.25, verbose=false);

julia> mean(ismissing.(genomes.allele_frequencies))
0.25

GBCore.simulategenomiceffects — Method

Simulate genomic effects

Simulate additive, dominance, and epistatic effects

Arguments

genomes: Genome struct includes the n entries x p loci-alleles combinations (p = l loci x a-1 alleles)

f_additive: proportion of the l loci with non-zero additive effects on the phenotype

f_dominance: proportion of the l*f_additive additive effects loci with additional dominance effects

f_epistasis: proportion of the l*f_additive additive effects loci with additional epistasis effects

Outputs

n x 3 matrix of additive, dominance and epistasis effects per entry

p x 3 matrix of additive, dominance and epistasis effects per locus-allele combination

Examples

julia> genomes::Genomes = simulategenomes(n=100, l=2_000, n_alleles=3, verbose=false);

julia> G, B = simulategenomiceffects(genomes=genomes, f_additive=0.05, f_dominance=0.75, f_epistasis=0.25);

julia> size.([G, B])
2-element Vector{Tuple{Int64, Int64}}:
 (100, 3)
 (4000, 3)

julia> sum(B .!= 0.0, dims=1)
1×3 Matrix{Int64}:
 200  75  50

Details

The additive, dominance, and epistasis allele effects share a common exponential distribution (λ=1) from which the mean of the effects (μ) are sampled, and the covariance matrix is derived (Σ = μ * μ'; where if det(Σ)≈0 then we iteratively add 1.00 to the diagonals until it becomes invertible or 10 iterations finishes and throws an error). The non-additive or epistasis allele effects were simulated by multiplying the allele frequencies of all possible unique pairs of epistasis alleles and their effects.

GBCore.simulatetrials — Method

Simulate trials

Arguments

genomes: Genome struct includes the n entries x p loci-alleles combinations (p = l loci x a-1 alleles)

f_add_dom_epi: n_traits x 3 numeric matrix of loci proportion with additive, dominance and epistasis effects, i.e. each column refers to:

f_additive: proportion of the l loci with non-zero additive effects on the phenotype

f_dominance: proportion of the l*f_additive additive effects loci with additional dominance effects

f_epistasis: proportion of the l*f_additive additive effects loci with additional epistasis effects

(default = [0.01 0.25 0.10; 0.05 0.50 0.25; 0.10 0.25 0.00])

n_years: Number of years (default = 2)

n_seasons: Number of seasons (default = 4)

n_harvests: Number of harvests (default = 2)

n_sites: Number of sites (default = 4)

n_replications: Number of replications (default = 2)

n_blocks: Number of blocks across the entire field layout (default = missing)

n_rows: Number of rows across the entire field layout (default = missing)

n_cols: Number of columns across the entire field layout (default = missing)

proportion_of_variance: 9 x n_traits numeric matrix of scaled/non-scaled proportion of variances allocated to genetic and environmental effects (default = missing; values will be sampled from a uniform distribution followed by a biased sample on the first row, i.e. additive effects row). The rows correspond to the variance allocated to:

additive genetic effects

dominance genetic effects

epistasis genetic effects

years effects

seasons effects

sites effects

environmental interactions

spatial interactions

GxE interactiions

seed: Randomisation seed (default = 42)

sparsity: Proportion of missing data (default = 0.0)

verbose: Show trials simulation progress bar? (default = true)

Outputs

Trials struct of simulated phenotype data

Vector of SimulatedEffects each corresponding to each trait-year-season-harvest-site-replication combination

Examples

julia> genomes::Genomes = simulategenomes(n=100, l=2_000, n_alleles=3, verbose=false);

julia> trials::Trials, vector_of_effects::Array{SimulatedEffects,1} = simulatetrials(genomes=genomes, sparsity=0.25, verbose=false);

julia> size(trials.phenotypes)
(12800, 3)

julia> size(trials.traits)
(3,)

julia> unique(trials.entries) == genomes.entries
true

julia> unique(trials.populations) == unique(genomes.populations)
true

julia> abs(mean(ismissing.(trials.phenotypes)) - 0.25) < 0.00001
true

GBCore.slice — Method

slice(
    genomes::Genomes; 
    idx_entries::Union{Nothing, Vector{Int64}} = nothing,
    idx_loci_alleles::Union{Nothing, Vector{Int64}} = nothing,
)::Genomes

Slice a Genomes struct by specifing indixes of entries and loci-allele combinations

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> sliced_genomes = slice(genomes, idx_entries=collect(1:10); idx_loci_alleles=collect(1:300));

julia> dimensions(sliced_genomes)
Dict{String, Int64} with 7 entries:
  "n_entries"      => 10
  "n_chr"          => 1
  "n_loci"         => 100
  "n_loci_alleles" => 300
  "n_populations"  => 1
  "n_missing"      => 0
  "max_n_alleles"  => 4

GBCore.slice — Method

slice(
    phenomes::Phenomes;
    idx_entries::Union{Nothing, Vector{Int64}} = nothing,
    idx_traits::Union{Nothing, Vector{Int64}} = nothing,
)::Phenomes

Slice a Phenomes struct by specifing indexes of entries and traits

Examples

julia> phenomes = Phenomes(n=10, t=3); phenomes.entries = string.("entry_", 1:10); phenomes.populations .= "pop_1"; phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = fill(0.0, 10,3);

julia> sliced_phenomes = slice(phenomes, idx_entries=collect(1:5); idx_traits=collect(2:3));

julia> dimensions(sliced_phenomes)
Dict{String, Int64} with 8 entries:
  "n_total"       => 10
  "n_zeroes"      => 10
  "n_nan"         => 0
  "n_entries"     => 5
  "n_traits"      => 2
  "n_inf"         => 0
  "n_populations" => 1
  "n_missing"     => 0

GBCore.slice — Method

slice(
    trials::Trials; 
    traits::Union{Nothing, Vector{String}} = nothing,
    populations::Union{Nothing, Vector{String}} = nothing,
    entries::Union{Nothing, Vector{String}} = nothing,
    years::Union{Nothing, Vector{String}} = nothing,
    harvests::Union{Nothing, Vector{String}} = nothing,
    seasons::Union{Nothing, Vector{String}} = nothing,
    sites::Union{Nothing, Vector{String}} = nothing,
    replications::Union{Nothing, Vector{String}} = nothing,
    blocks::Union{Nothing, Vector{String}} = nothing,
    rows::Union{Nothing, Vector{String}} = nothing,
    cols::Union{Nothing, Vector{String}} = nothing,
)::Trials

Slice a Trials struct by specifing indexes of entries and traits

Examples

julia> trials, _ = simulatetrials(genomes = simulategenomes(verbose=false), verbose=false);

julia> sliced_trials = slice(trials, traits=trials.traits[2:3], years=[unique(trials.years)[1]], seasons=unique(trials.seasons)[2:3]);

julia> dimensions(sliced_trials)
Dict{String, Int64} with 16 entries:
  "n_zeroes"       => 0
  "n_harvests"     => 2
  "n_nan"          => 0
  "n_entries"      => 100
  "n_traits"       => 2
  "n_seasons"      => 2
  "n_blocks"       => 10
  "n_rows"         => 10
  "n_missing"      => 0
  "n_inf"          => 0
  "n_total"        => 6400
  "n_replications" => 2
  "n_years"        => 1
  "n_sites"        => 4
  "n_cols"         => 20
  "n_populations"  => 1

GBCore.summarise — Method

summarise(cvs::Vector{CV})::Tuple{DataFrame,DataFrame}

Summarise a vector of CV structs into:

a data frame of mean metrics, and

a data frame of mean and standard deviation of phenotype predictions per entry, trait and model.

Examples

julia> fit_1 = Fit(n=1, l=2); fit_1.metrics = Dict("cor" => 0.0, "rmse" => 1.0); fit_1.trait = "trait_1";

julia> cv_1 = CV("replication_1", "fold_1", fit_1, ["population_1"], ["entry_1"], [0.0], [0.0], fit_1.metrics);

julia> fit_2 = Fit(n=1, l=2); fit_2.metrics = Dict("cor" => 1.0, "rmse" => 0.0); fit_2.trait = "trait_1";

julia> cv_2 = CV("replication_2", "fold_2", fit_2, ["population_2"], ["entry_2"], [0.0], [0.0], fit_2.metrics);

julia> cvs = [cv_1, cv_2];

julia> df_summary, df_summary_per_entry = summarise(cvs);

julia> size(df_summary)
(2, 6)

julia> size(df_summary_per_entry)
(2, 8)

GBCore.tabularise — Function

tabularise(trials::Trials)::DataFrame

Export the Trials structs into a DataFrames.DataFrame struct

Examples

julia> fit = Fit(n=100, l=10_000); fit.b_hat = rand(10_000); fit.model="some_model"; fit.trait="some_trait"; 

julia> fit.metrics = Dict("cor" => rand(), "rmse" => rand()); fit.populations .= "pop_1";

julia> df = tabularise(fit);

julia> size(df)
(10000, 6)

GBCore.tabularise — Method

tabularise(phenomes::Phenomes)::DataFrame

Export the Phenomes structs into a DataFrames.DataFrame struct

Examples

julia> phenomes = Phenomes(n=10, t=3); phenomes.entries = string.("entry_", 1:10); phenomes.populations .= "pop_1"; phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = fill(0.0, 10,3);

julia> tabularise(phenomes)
10×6 DataFrame
 Row │ id     entries   populations  A         B         C        
     │ Int64  String    String       Float64?  Float64?  Float64? 
─────┼────────────────────────────────────────────────────────────
   1 │     1  entry_1   pop_1             0.0       0.0       0.0
   2 │     2  entry_2   pop_1             0.0       0.0       0.0
   3 │     3  entry_3   pop_1             0.0       0.0       0.0
   4 │     4  entry_4   pop_1             0.0       0.0       0.0
   5 │     5  entry_5   pop_1             0.0       0.0       0.0
   6 │     6  entry_6   pop_1             0.0       0.0       0.0
   7 │     7  entry_7   pop_1             0.0       0.0       0.0
   8 │     8  entry_8   pop_1             0.0       0.0       0.0
   9 │     9  entry_9   pop_1             0.0       0.0       0.0
  10 │    10  entry_10  pop_1             0.0       0.0       0.0

GBCore.tabularise — Method

tabularise(trials::Trials)::DataFrame

Export the Trials structs into a DataFrames.DataFrame struct

Examples

julia> trials, _ = simulatetrials(genomes = simulategenomes(verbose=false), verbose=false);

julia> df = tabularise(trials);

julia> size(df)
(12800, 14)

GBCore.tabularise — Method

tabularise(cvs::Vector{CV})::Tuple{DataFrame,DataFrame}

Export a vector of CV structs into data frames of metrics across entries and per validation entry

Examples

julia> fit_1 = Fit(n=1, l=2); fit_1.metrics = Dict("cor" => 0.0, "rmse" => 1.0); fit_1.trait = "trait_1";

julia> cv_1 = CV("replication_1", "fold_1", fit_1, ["population_1"], ["entry_1"], [0.0], [0.0], fit_1.metrics);

julia> fit_2 = Fit(n=1, l=2); fit_2.metrics = Dict("cor" => 1.0, "rmse" => 0.0); fit_2.trait = "trait_1";

julia> cv_2 = CV("replication_2", "fold_2", fit_2, ["population_2"], ["entry_2"], [0.0], [0.0], fit_2.metrics);

julia> cvs = [cv_1, cv_2];

julia> df_across_entries, df_per_entry = tabularise(cvs);

julia> names(df_across_entries)
8-element Vector{String}:
 "training_population"
 "validation_population"
 "trait"
 "model"
 "replication"
 "fold"
 "cor"
 "rmse"

julia> df_across_entries[!, [:cor, :rmse]]
2×2 DataFrame
 Row │ cor      rmse    
     │ Float64  Float64 
─────┼──────────────────
   1 │     0.0      1.0
   2 │     1.0      0.0

julia> names(df_per_entry)
9-element Vector{String}:
 "training_population"
 "validation_population"
 "entry"
 "trait"
 "model"
 "replication"
 "fold"
 "y_true"
 "y_pred"

julia> df_per_entry[!, [:entry, :y_true, :y_pred]]
2×3 DataFrame
 Row │ entry    y_true   y_pred  
     │ String   Float64  Float64 
─────┼───────────────────────────
   1 │ entry_1      0.0      0.0
   2 │ entry_2      0.0      0.0

GBCore.trialsmodelsfomulae! — Method

trialsmodelsfomulae!(df::DataFrame; trait::String, max_levels::Int64 = 100)::Vector{String}

Define formulae for the mixed models to fit on the tabularised Trials struct. - appends interaction effects intto df - returns: + a vector of formulae as strings + a vector of the total number of non-entry factor levels

Examples

julia> trials, _simulated_effects = simulatetrials(genomes = simulategenomes(verbose=false), verbose=false);

julia> df = tabularise(trials);

julia> size(df)
(12800, 14)

julia> formulae, n_levels = trialsmodelsfomulae!(df, trait="trait_1");

julia> size(df)
(12800, 134)

julia> length(formulae)
76

julia> sum(n_levels .== sort(n_levels))
76

GBCore.@string2formula — Macro

string2formula(x)

Macro to Meta.parse a string into a formula. TODO: Refactor to make more idiomatic.

GBCore.@stringevaluation — Macro

stringevaluation(x)

Macro to Meta.parse a string of formula.